# code/go/api/restapi/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package restapi

import (
	"fmt"
	"net/url"
	"time"
)

func (s *ProblemStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

type ApiKeyAuth struct {
	APIKey string
	Roles  []string
}

// GetAPIKey returns the value of APIKey.
func (s *ApiKeyAuth) GetAPIKey() string {
	return s.APIKey
}

// GetRoles returns the value of Roles.
func (s *ApiKeyAuth) GetRoles() []string {
	return s.Roles
}

// SetAPIKey sets the value of APIKey.
func (s *ApiKeyAuth) SetAPIKey(val string) {
	s.APIKey = val
}

// SetRoles sets the value of Roles.
func (s *ApiKeyAuth) SetRoles(val []string) {
	s.Roles = val
}

// HealthOK is response for Health operation.
type HealthOK struct{}

// NewOptInt32 returns new OptInt32 with value set to v.
func NewOptInt32(v int32) OptInt32 {
	return OptInt32{
		Value: v,
		Set:   true,
	}
}

// OptInt32 is optional int32.
type OptInt32 struct {
	Value int32
	Set   bool
}

// IsSet returns true if OptInt32 was set.
func (o OptInt32) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptInt32) Reset() {
	var v int32
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptInt32) SetTo(v int32) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptInt32) Get() (v int32, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptInt32) Or(d int32) int32 {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptURI returns new OptURI with value set to v.
func NewOptURI(v url.URL) OptURI {
	return OptURI{
		Value: v,
		Set:   true,
	}
}

// OptURI is optional url.URL.
type OptURI struct {
	Value url.URL
	Set   bool
}

// IsSet returns true if OptURI was set.
func (o OptURI) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptURI) Reset() {
	var v url.URL
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptURI) SetTo(v url.URL) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptURI) Get() (v url.URL, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptURI) Or(d url.URL) url.URL {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/Problem
type Problem struct {
	// An absolute URI that identifies the problem type.  When dereferenced,
	// it SHOULD provide human-readable documentation for the problem type
	// (e.g., using HTML).
	Type OptURI `json:"type"`
	// A short, summary of the problem type. Written in english and readable
	// for engineers (usually not suited for non technical stakeholders and
	// not localized); example: Service Unavailable.
	Title OptString `json:"title"`
	// The HTTP status code generated by the origin server for this occurrence
	// of the problem.
	Status OptInt32 `json:"status"`
	// A human readable explanation specific to this occurrence of the
	// problem.
	Detail OptString `json:"detail"`
	// An absolute URI that identifies the specific occurrence of the problem.
	// It may or may not yield further information if dereferenced.
	Instance OptURI `json:"instance"`
	// The error code generated by the origin server for this occurrence
	// of the problem.
	ErrorCode OptString `json:"error_code"`
}

// GetType returns the value of Type.
func (s *Problem) GetType() OptURI {
	return s.Type
}

// GetTitle returns the value of Title.
func (s *Problem) GetTitle() OptString {
	return s.Title
}

// GetStatus returns the value of Status.
func (s *Problem) GetStatus() OptInt32 {
	return s.Status
}

// GetDetail returns the value of Detail.
func (s *Problem) GetDetail() OptString {
	return s.Detail
}

// GetInstance returns the value of Instance.
func (s *Problem) GetInstance() OptURI {
	return s.Instance
}

// GetErrorCode returns the value of ErrorCode.
func (s *Problem) GetErrorCode() OptString {
	return s.ErrorCode
}

// SetType sets the value of Type.
func (s *Problem) SetType(val OptURI) {
	s.Type = val
}

// SetTitle sets the value of Title.
func (s *Problem) SetTitle(val OptString) {
	s.Title = val
}

// SetStatus sets the value of Status.
func (s *Problem) SetStatus(val OptInt32) {
	s.Status = val
}

// SetDetail sets the value of Detail.
func (s *Problem) SetDetail(val OptString) {
	s.Detail = val
}

// SetInstance sets the value of Instance.
func (s *Problem) SetInstance(val OptURI) {
	s.Instance = val
}

// SetErrorCode sets the value of ErrorCode.
func (s *Problem) SetErrorCode(val OptString) {
	s.ErrorCode = val
}

// ProblemStatusCode wraps Problem with StatusCode.
type ProblemStatusCode struct {
	StatusCode int
	Response   Problem
}

// GetStatusCode returns the value of StatusCode.
func (s *ProblemStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ProblemStatusCode) GetResponse() Problem {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ProblemStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ProblemStatusCode) SetResponse(val Problem) {
	s.Response = val
}

// Ref: #/components/responses/400
type R400 struct{}

func (*R400) recordTollEventRes() {}

// Ref: #/components/responses/401
type R401 struct{}

func (*R401) recordTollEventRes() {}

// Ref: #/components/responses/403
type R403 struct{}

func (*R403) recordTollEventRes() {}

// RecordTollEventNoContent is response for RecordTollEvent operation.
type RecordTollEventNoContent struct{}

func (*RecordTollEventNoContent) recordTollEventRes() {}

// Ref: #/components/schemas/TollEvent
type TollEvent struct {
	VehicleId string    `json:"vehicleId"`
	StationId string    `json:"stationId"`
	Timestamp time.Time `json:"timestamp"`
}

// GetVehicleId returns the value of VehicleId.
func (s *TollEvent) GetVehicleId() string {
	return s.VehicleId
}

// GetStationId returns the value of StationId.
func (s *TollEvent) GetStationId() string {
	return s.StationId
}

// GetTimestamp returns the value of Timestamp.
func (s *TollEvent) GetTimestamp() time.Time {
	return s.Timestamp
}

// SetVehicleId sets the value of VehicleId.
func (s *TollEvent) SetVehicleId(val string) {
	s.VehicleId = val
}

// SetStationId sets the value of StationId.
func (s *TollEvent) SetStationId(val string) {
	s.StationId = val
}

// SetTimestamp sets the value of Timestamp.
func (s *TollEvent) SetTimestamp(val time.Time) {
	s.Timestamp = val
}