# code/spec/openapi.yaml
---
openapi: 3.1.0

info:
  title: Toll API
  version: '1.0.0'
  description: |
    The Toll API provides real-time updates about toll reporting events

servers:
  - url: /api/v1/

security:
  - {}
  - ApiKeyAuth: []

paths:
  /health:
    get:
      description: Returns OK.
      summary: Returns OK when service is running.
      operationId: Health
      security:
        - {}
      responses:
        '200':
          description: OK response
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

  /toll-events:
    post:
      summary: Record toll event.
      description: Stores toll event log to event database.
      operationId: RecordTollEvent
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        description: Toll event to record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TollEvent'
      responses:
        204:
          description: Toll event created
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

parameters:
  Id:
    in: path
    name: id
    required: true
    schema:
      type: string
    description: Required ID

  UUID:
    in: path
    name: id
    required: true
    schema:
      type: string
      format: uuid
    description: Required ID

components:
  schemas:
    TollEvent:
      type: object
      required:
        - vehicleId
        - stationId
        - timestamp
      properties:
        vehicleId:
          type: string
          example: 'ABC1234'
        stationId:
          type: string
          example: 'TOLL-GANTRY-17'
        timestamp:
          type: string
          format: date-time
          example: '2025-02-15T14:32:10Z'

    Error:
      type: object
      properties:
        code:
          type: string
          enum:
            - internal_error
            - not_found
          example: unauthorized
        message:
          type: string
          example: Unauthorized access

    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: |
            An absolute URI that identifies the problem type.  When dereferenced,
            it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          default: 'about:blank'
          example: 'https://zalando.github.io/problem/constraint-violation'
        title:
          type: string
          description: |
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized); example: Service Unavailable
        status:
          type: integer
          format: int32
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          minimum: 100
          maximum: 600
          exclusiveMaximum: true
          example: 503
        detail:
          type: string
          description: |
            A human readable explanation specific to this occurrence of the
            problem.
          example: Connection to database timed out
        instance:
          type: string
          format: uri
          description: |
            An absolute URI that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced.
        error_code:
          type: string
          description: |
            The error code generated by the origin server for this occurrence
            of the problem.

  responses:
    '400':
      description: Request is badly formatted.
      schema:
        $ref: '#/definitions/Error'
      examples:
        application/json:
          code: bad_request
          message: Request is badly formatted.

    '401':
      description: Unauthorized Access
      schema:
        $ref: '#/definitions/Error'
      examples:
        application/json:
          code: unauthorized
          message: Unauthorized Access

    '403':
      description: Forbiden
      schema:
        $ref: '#/definitions/Error'
      examples:
        application/json:
          code: forbiden
          message: You do not have permissions to do this

    '404':
      description: Required entity cannot be found
      schema:
        $ref: '#/definitions/Error'
      examples:
        application/json:
          code: not_found
          message: Required entity cannot be found

    '500':
      description: Internal server error
      schema:
        $ref: '#/definitions/Error'
      examples:
        application/json:
          title: Something went wrong
          status: 500
          detail: Something went wrong.

    '503':
      description: Service unavailable
      schema:
        $ref: '#/definitions/Error'
      examples:
        application/json:
          code: service_unavailable
          message: Service unavailable

    error:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: internalError
            message: Internal server error

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key